+++ src/main/java/database/dialect/MySQLDialect.java
@@ -1,7 +1,5 @@
 package database.dialect;
 
-import database.mapping.SqlTypes;
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Types;
