+++ src/main/java/database/mapping/rowmapper/MultiRowMapper.java
@@ -1,4 +1,4 @@
-package database.mapping;
+package database.mapping.rowmapper;
 
 import database.dialect.MySQLDialect;
 import jdbc.RowMapper;
