+++ src/main/java/database/mapping/rowmapper/RowMapMerger.java
@@ -1,4 +1,8 @@
-package database.mapping;
+package database.mapping.rowmapper;
+
+import database.mapping.Association;
+import database.mapping.EntityMetadata;
+import database.mapping.EntityMetadataFactory;
 
 import java.lang.reflect.Field;
 import java.util.List;
